requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
youtube-transcript-api>=1.0.0
gspread>=6.0.0
pytrends>=4.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
requests>=2.31.0
notebooklm-py>=0.3.0
bcrypt>=4.0.0
cryptography>=41.0.0
slowapi>=0.1.9
httpx>=0.27.0
python-dotenv>=1.0.0
